Change unknown_packet into login_packet, corrected some bugs

+++ src/protocol/streams/output_stream.cpp
@@ -20,7 +20,11 @@ void output_stream::writeString(std::string& text) {
 }
 
 void output_stream::writePosition(position pos) {
-    writeLong(((pos.x & 0x3FFFFFF) << 38) | ((pos.z & 0x3FFFFFF) << 12) | (pos.y & 0xFFF));
+    writeLong(
+        ((((uint64_t)pos.x) & 0x3FFFFFF) << 38) | 
+        ((((uint64_t)pos.z) & 0x3FFFFFF) << 12) | 
+        (((uint64_t)pos.y) & 0xFFF)
+    );
 }
 
 void output_stream::writeByte(uint8_t send) {
